This commits includes other changes and improvements requested by Gianluca. In particular, software related to CRT Hit transformation has beend moved from the namespace dedicated to the matching, to a new one dedicated to data related tools
