+++ icaruscode/CRT/CRTUtils/RecoUtils.h
@@ -43,42 +43,94 @@ namespace detinfo { class DetectorClocksData; }
 #include "TTree.h"
 
 
+namespace icarus::crt::dataTools{
+
+struct ModuleCenter
+{
+  double X;
+  double Y;
+  double Z;
+};
+
+struct AffineTrans
+{
+  double A11;
+  double A12;
+  double A21;
+  double A22;
+  double B1;
+  double B2;
+  double Accuracy;
+  double N;
+};
+
+using FebIndex_t = int;
+
+using TopCRTCorrectionMap = std::map<FebIndex_t, AffineTrans>;
+
+struct TopCRTTransformations
+{
+  TopCRTCorrectionMap EE;
+  TopCRTCorrectionMap EW;
+  TopCRTCorrectionMap EastCC;
+  TopCRTCorrectionMap WE;
+  TopCRTCorrectionMap WW;
+  TopCRTCorrectionMap WestCC;
+};
+
+using TopCRTCentersMap = std::map<FebIndex_t, ModuleCenter>;
+
+/// The transformed CRT Hits are in cm
+using TransformedCRTHit = std::pair<double, double>;
+
+/// This function loads the Top CRT modules centers.
+TopCRTCentersMap LoadTopCRTCenters();
+
+/// This function performs the affine transformation of the CRT hit points.
+/// The AffineTransformation requires input variables in cm.
+TransformedCRTHit AffineTransformation(double DX, double DZ, AffineTrans affine);
+
+/// This functions loads the Affine Transformation TXT files.
+TopCRTTransformations LoadTopCRTTransformations();
+
+}
+
 namespace RecoUtils{
 
-  // Returns the geant4 ID which contributes the most to a single reco hit.  
-  // The matching method looks for true particle which deposits the most true energy in the reco hit.  
-  // If rollup_unsaved_ids is set to true, any unsaved daughter than 
-  // contributed energy to the hit has its energy included in its closest ancestor that was saved.
+// Returns the geant4 ID which contributes the most to a single reco hit.  
+// The matching method looks for true particle which deposits the most true energy in the reco hit.  
+// If rollup_unsaved_ids is set to true, any unsaved daughter than 
+// contributed energy to the hit has its energy included in its closest ancestor that was saved.
 
-  int TrueParticleID(detinfo::DetectorClocksData const& clockData, const art::Ptr<recob::Hit> hit, bool rollup_unsaved_ids=1); 
+int TrueParticleID(detinfo::DetectorClocksData const& clockData, const art::Ptr<recob::Hit> hit, bool rollup_unsaved_ids=1); 
   
-  // Returns the geant4 ID which contributes the most to the vector of hits.  
-  // The matching method looks for which true particle deposits the most true energy in the reco hits
+// Returns the geant4 ID which contributes the most to the vector of hits.  
+// The matching method looks for which true particle deposits the most true energy in the reco hits
 
-  int TrueParticleIDFromTotalTrueEnergy(detinfo::DetectorClocksData const& clockData, const std::vector<art::Ptr<recob::Hit> >& hits, bool rollup_unsaved_ids=1); 
+int TrueParticleIDFromTotalTrueEnergy(detinfo::DetectorClocksData const& clockData, const std::vector<art::Ptr<recob::Hit> >& hits, bool rollup_unsaved_ids=1); 
 
-  // Returns the geant4 ID which contributes the most to the vector of hits.  
-  // The matching method looks for which true particle contributes the most reconstructed charge to the hit selection 
-  // (the reco charge of each hit is correlated with each maximally contributing true particle and summed)
+// Returns the geant4 ID which contributes the most to the vector of hits.  
+// The matching method looks for which true particle contributes the most reconstructed charge to the hit selection 
+// (the reco charge of each hit is correlated with each maximally contributing true particle and summed)
   
-  int TrueParticleIDFromTotalRecoCharge(detinfo::DetectorClocksData const& clockData, const std::vector<art::Ptr<recob::Hit> >& hits, bool rollup_unsaved_ids=1); 
+int TrueParticleIDFromTotalRecoCharge(detinfo::DetectorClocksData const& clockData, const std::vector<art::Ptr<recob::Hit> >& hits, bool rollup_unsaved_ids=1); 
 
-  // Returns the geant4 ID which contributes the most to the vector of hits.  
-  // The matching method looks for which true particle maximally contributes to the most reco hits
+// Returns the geant4 ID which contributes the most to the vector of hits.  
+// The matching method looks for which true particle maximally contributes to the most reco hits
   
-  int TrueParticleIDFromTotalRecoHits(detinfo::DetectorClocksData const& clockData, const std::vector<art::Ptr<recob::Hit> >& hits, bool rollup_unsaved_ids=1); 
+int TrueParticleIDFromTotalRecoHits(detinfo::DetectorClocksData const& clockData, const std::vector<art::Ptr<recob::Hit> >& hits, bool rollup_unsaved_ids=1); 
 
-  // Checks if a position is within any of the TPCs in the geometry (user can define some distance buffer from the TPC walls)
+// Checks if a position is within any of the TPCs in the geometry (user can define some distance buffer from the TPC walls)
   
-  bool IsInsideTPC(TVector3 position, double distance_buffer); 
+bool IsInsideTPC(TVector3 position, double distance_buffer); 
 
-  // Calculates the total length of a recob::track by summing up the distances between adjacent traj. points
+// Calculates the total length of a recob::track by summing up the distances between adjacent traj. points
 
-  double CalculateTrackLength(const art::Ptr<recob::Track> track); 
+double CalculateTrackLength(const art::Ptr<recob::Track> track); 
   
-  std::map<int, std::vector<std::pair<geo::WireID, const sim::IDE*>>> PrepSimChannels(const std::vector<art::Ptr<sim::SimChannel>> &simchannels, const geo::GeometryCore &geo);
+std::map<int, std::vector<std::pair<geo::WireID, const sim::IDE*>>> PrepSimChannels(const std::vector<art::Ptr<sim::SimChannel>> &simchannels, const geo::GeometryCore &geo);
 
-  std::map<int, std::vector<art::Ptr<recob::Hit>>> buildTrackIDtoHitsMap(const std::vector<art::Ptr<recob::Hit>> &allHits, const detinfo::DetectorClocksData &clockData, const cheat::BackTrackerService &backtracker);
+std::map<int, std::vector<art::Ptr<recob::Hit>>> buildTrackIDtoHitsMap(const std::vector<art::Ptr<recob::Hit>> &allHits, const detinfo::DetectorClocksData &clockData, const cheat::BackTrackerService &backtracker);
 }
 
 #endif
